+++ src/model/solver/cone_class.hpp
@@ -17,23 +17,23 @@ struct Cone {
    ~Cone() = default;
 
    // constructors
-   Cone(const vm::Vec3<float> p0, const vm::Vec3<float> p1, const float r0,
-        const float r1, const long long fiber_idx, const long long cone_idx)
+   Cone(const vm::Vec3<double> p0, const vm::Vec3<double> p1, const double r0,
+        const double r1, const long long fiber_idx, const long long cone_idx)
        : p0(p0), p1(p1), r0(r0), r1(r1), r(r0 > r1 ? r0 : r1),
          fiber_idx(fiber_idx), cone_idx(cone_idx){};
 
-   aabb::AABB<float, 3> aabb() const;
+   aabb::AABB<double, 3> aabb() const;
 
    // colliding solving functions
    bool CollideWith(const Cone cone) const;
-   std::tuple<vm::Vec3<float>, vm::Vec3<float>>
+   std::tuple<vm::Vec3<double>, vm::Vec3<double>>
    MinDistanceVectorCones(const Cone cone) const;
-   std::tuple<vm::Vec3<float>, vm::Vec3<float>, vm::Vec3<float>,
-              vm::Vec3<float>>
+   std::tuple<vm::Vec3<double>, vm::Vec3<double>, vm::Vec3<double>,
+              vm::Vec3<double>>
    PushConesApart(const Cone cone) const;
 
-   vm::Vec3<float> p0, p1;
-   float r0, r1, r; // TODO: check if r0, r1 are needed
+   vm::Vec3<double> p0, p1;
+   double r0, r1, r; // TODO: check if r0, r1 are needed
    size_t fiber_idx, cone_idx;
 };
 } // namespace object
